feat(fish): update config.fish

+++ fish/dot-config/fish/config.fish
@@ -1,28 +1,28 @@
-# if status is-interactive
-#     # Commands to run in interactive sessions can go here
-# end
+if status is-interactive
+    # Commands to run in interactive sessions can go here
 
-# Tmux
-abbr t 'sesh connect $(sesh list | fzf)'
-# abbr t tmux
-abbr tc 'tmux attach'
-abbr ta 'tmux attach -t'
-abbr tad 'tmux attach -d -t'
-abbr ts 'tmux new -s'
-abbr tl 'tmux ls'
-abbr tk 'tmux kill-session -t'
+    # Tmux
+    abbr t 'sesh connect $(sesh list | fzf)'
+    # abbr t tmux
+    abbr tc 'tmux attach'
+    abbr ta 'tmux attach -t'
+    abbr tad 'tmux attach -d -t'
+    abbr ts 'tmux new -s'
+    abbr tl 'tmux ls'
+    abbr tk 'tmux kill-session -t'
 
-# Editor
-abbr vim nvim
-abbr vi nvim
-abbr v nvim
+    # Editor
+    abbr vim nvim
+    abbr vi nvim
+    abbr v nvim
 
-# Dev
-abbr gg lazygit
-abbr gb 'git checkout -b'
-abbr gc 'git commit'
-abbr gp 'git pull'
-abbr gP 'git push'
+    # Dev
+    abbr gg lazygit
+    abbr gb 'git checkout -b'
+    abbr gc 'git commit'
+    abbr gp 'git pull'
+    abbr gP 'git push'
 
-starship init fish | source
-zoxide init fish | source
+    starship init fish | source
+    zoxide init fish | source
+end
